Double, Long, Character types changes

+++ dfd/dfdAlgorithm/src/fdiscovery/approach/runner/DFDMiner.java
@@ -116,7 +116,7 @@ public class DFDMiner extends Miner implements Runnable {
 
 	private void writeOutputSuccessful(String outputFile, long time, String inputFileName) {
 
-		String timeString = (time != -1) ? String.format("%.1f", (double) (time) / 1000) : "-1";
+		String timeString = (time != -1) ? String.format("%.1f", Double.valueOf((double) (time) / 1000)) : "-1";
 		StringBuilder outputBuilder = new StringBuilder();
 		if (!inputFileName.isEmpty()) {
 			outputBuilder.append(String.format("%s\t", inputFileName));
